+++ packages/block/typedoc.js
@@ -1,12 +1,10 @@
 module.exports = {
-  inputFiles: ['src/*.ts'],
-  mode: 'library',
+  entryPoints: ['src'],
   out: 'docs',
   plugin: 'typedoc-plugin-markdown',
   readme: 'none',
   gitRevision: 'master',
   exclude: 'test/**/*.ts',
-  excludeNotExported: true,
   excludePrivate: true,
   excludeProtected: true,
 }