+++ packages/blockchain/typedoc.js
@@ -1,6 +1,5 @@
 module.exports = {
-  inputFiles: ['src/*.ts'],
-  mode: 'library',
+  entryPoints: ['src'],
   out: 'docs',
   plugin: 'typedoc-plugin-markdown',
   readme: 'none',
@@ -10,7 +9,6 @@ module.exports = {
     'src/db/**',
     'src/clique.ts',
   ],
-  excludeNotExported: true,
   excludePrivate: true,
   excludeProtected: true,
 }